Move integration tests from unit-testing into integration-testing (#1370)
Python unit tests currently comprise integration tests that are slow or
impossible to run OOTB from a desktop.
This PR solves the issue by:
- moving integration tests from the `units` hierarchy to the
`integration` one
- adding a `python-integration` job to the CI that runs tests under
`integration`
