+++ tests/unit/reconcile/connectors/test_snowflake.py
@@ -68,7 +68,7 @@ def mock_malformed_private_key_secret(scope, key):
 
 
 def mock_no_auth_key_secret(scope, key):
-    if key == 'pem_private_key' or key == 'sfPassword':
+    if key in {'pem_private_key', 'sfPassword'}:
         raise NotFound("Secret not found")
     return mock_secret(scope, key)
 
@@ -89,8 +89,8 @@ def test_get_jdbc_url_happy():
     # initial setup
     engine, spark, ws, scope = initial_setup()
     # create object for SnowflakeDataSource
-    ds = SnowflakeDataSource(engine, spark, ws, scope)
-    url = ds.get_jdbc_url
+    dfds = SnowflakeDataSource(engine, spark, ws, scope)
+    url = dfds.get_jdbc_url
     # Assert that the URL is generated correctly
     assert url == (
         "jdbc:snowflake://my_account.snowflakecomputing.com"
@@ -105,8 +105,8 @@ def test_get_jdbc_url_fail():
     engine, spark, ws, scope = initial_setup()
     ws.secrets.get_secret.side_effect = mock_secret
     # create object for SnowflakeDataSource
-    ds = SnowflakeDataSource(engine, spark, ws, scope)
-    url = ds.get_jdbc_url
+    dfds = SnowflakeDataSource(engine, spark, ws, scope)
+    url = dfds.get_jdbc_url
     # Assert that the URL is generated correctly
     assert url == (
         "jdbc:snowflake://my_account.snowflakecomputing.com"
@@ -121,23 +121,15 @@ def test_read_data_with_out_options():
     engine, spark, ws, scope = initial_setup()
 
     # create object for SnowflakeDataSource
-    ds = SnowflakeDataSource(engine, spark, ws, scope)
+    dfds = SnowflakeDataSource(engine, spark, ws, scope)
     # Create a Tables configuration object with no JDBC reader options
     table_conf = Table(
         source_name="supplier",
         target_name="supplier",
-        jdbc_reader_options=None,
-        join_columns=None,
-        select_columns=None,
-        drop_columns=None,
-        column_mapping=None,
-        transformations=None,
-        column_thresholds=None,
-        filters=None,
     )
 
     # Call the read_data method with the Tables configuration
-    ds.read_data("org", "data", "employee", "select 1 from :tbl", table_conf.jdbc_reader_options)
+    dfds.read_data("org", "data", "employee", "select 1 from :tbl", table_conf.jdbc_reader_options)
 
     # spark assertions
     spark.read.format.assert_called_with("snowflake")
@@ -159,7 +151,7 @@ def test_read_data_with_options():
     engine, spark, ws, scope = initial_setup()
 
     # create object for SnowflakeDataSource
-    ds = SnowflakeDataSource(engine, spark, ws, scope)
+    dfds = SnowflakeDataSource(engine, spark, ws, scope)
     # Create a Tables configuration object with JDBC reader options
     table_conf = Table(
         source_name="supplier",
@@ -177,7 +169,7 @@ def test_read_data_with_options():
     )
 
     # Call the read_data method with the Tables configuration
-    ds.read_data("org", "data", "employee", "select 1 from :tbl", table_conf.jdbc_reader_options)
+    dfds.read_data("org", "data", "employee", "select 1 from :tbl", table_conf.jdbc_reader_options)
 
     # spark assertions
     spark.read.format.assert_called_with("jdbc")
@@ -199,9 +191,9 @@ def test_get_schema():
     engine, spark, ws, scope = initial_setup()
     # Mocking get secret method to return the required values
     # create object for SnowflakeDataSource
-    ds = SnowflakeDataSource(engine, spark, ws, scope)
+    dfds = SnowflakeDataSource(engine, spark, ws, scope)
     # call test method
-    ds.get_schema("catalog", "schema", "supplier")
+    dfds.get_schema("catalog", "schema", "supplier")
     # spark assertions
     spark.read.format.assert_called_with("snowflake")
     spark.read.format().option.assert_called_with(
@@ -231,7 +223,7 @@ def test_get_schema():
 def test_read_data_exception_handling():
     # initial setup
     engine, spark, ws, scope = initial_setup()
-    ds = SnowflakeDataSource(engine, spark, ws, scope)
+    dfds = SnowflakeDataSource(engine, spark, ws, scope)
     # Create a Tables configuration object
     table_conf = Table(
         source_name="supplier",
@@ -253,14 +245,14 @@ def test_read_data_exception_handling():
         DataSourceRuntimeException,
         match="Runtime exception occurred while fetching data using select 1 from org.data.employee : Test Exception",
     ):
-        ds.read_data("org", "data", "employee", "select 1 from :tbl", table_conf.jdbc_reader_options)
+        dfds.read_data("org", "data", "employee", "select 1 from :tbl", table_conf.jdbc_reader_options)
 
 
 def test_get_schema_exception_handling():
     # initial setup
     engine, spark, ws, scope = initial_setup()
 
-    ds = SnowflakeDataSource(engine, spark, ws, scope)
+    dfds = SnowflakeDataSource(engine, spark, ws, scope)
 
     spark.read.format().option().options().load.side_effect = RuntimeError("Test Exception")
 
@@ -275,15 +267,15 @@ def test_get_schema_exception_handling():
         "where lower\\(table_name\\)='supplier' and table_schema = 'SCHEMA' order by ordinal_position : Test "
         "Exception",
     ):
-        ds.get_schema("catalog", "schema", "supplier")
+        dfds.get_schema("catalog", "schema", "supplier")
 
 
 def test_read_data_with_out_options_private_key():
     engine, spark, ws, scope = initial_setup()
     ws.secrets.get_secret.side_effect = mock_private_key_secret
-    ds = SnowflakeDataSource(engine, spark, ws, scope)
+    dfds = SnowflakeDataSource(engine, spark, ws, scope)
     table_conf = Table(source_name="supplier", target_name="supplier")
-    ds.read_data("org", "data", "employee", "select 1 from :tbl", table_conf.jdbc_reader_options)
+    dfds.read_data("org", "data", "employee", "select 1 from :tbl", table_conf.jdbc_reader_options)
     spark.read.format.assert_called_with("snowflake")
     spark.read.format().option.assert_called_with("dbtable", "(select 1 from org.data.employee) as tmp")
     expected_options = {
@@ -303,18 +295,18 @@ def test_read_data_with_out_options_private_key():
 def test_read_data_with_out_options_malformed_private_key():
     engine, spark, ws, scope = initial_setup()
     ws.secrets.get_secret.side_effect = mock_malformed_private_key_secret
-    ds = SnowflakeDataSource(engine, spark, ws, scope)
+    dfds = SnowflakeDataSource(engine, spark, ws, scope)
     table_conf = Table(source_name="supplier", target_name="supplier")
     with pytest.raises(InvalidSnowflakePemPrivateKey, match="Failed to load or process the provided PEM private key."):
-        ds.read_data("org", "data", "employee", "select 1 from :tbl", table_conf.jdbc_reader_options)
+        dfds.read_data("org", "data", "employee", "select 1 from :tbl", table_conf.jdbc_reader_options)
 
 
 def test_read_data_with_out_any_auth():
     engine, spark, ws, scope = initial_setup()
     ws.secrets.get_secret.side_effect = mock_no_auth_key_secret
-    ds = SnowflakeDataSource(engine, spark, ws, scope)
+    dfds = SnowflakeDataSource(engine, spark, ws, scope)
     table_conf = Table(source_name="supplier", target_name="supplier")
     with pytest.raises(
         NotFound, match='sfPassword and pem_private_key not found. Either one is required for snowflake auth.'
     ):
-        ds.read_data("org", "data", "employee", "select 1 from :tbl", table_conf.jdbc_reader_options)
+        dfds.read_data("org", "data", "employee", "select 1 from :tbl", table_conf.jdbc_reader_options)
